+++ src/problems/index.ts
@@ -15,12 +15,18 @@ export {
   uciToSan,
   sanLineFromUci,
   isSolutionMove,
+  gradeLine,
   engineLineSummaries,
   formatSanLine,
 } from './solve'
+export type { LineDeviation, LineGrade } from './solve'
 export {
   readCheckComment,
+  wrongLineVerdict,
+  buildingLineNotice,
+  tryAgainNotice,
   wrongMoveExplanation,
+  wrongDefenceExplanation,
   fixReminder,
   curatedMotifNames,
   solvedMessage,
@@ -29,4 +35,9 @@ export {
   oneLineAttemptNotice,
   exploreHint,
 } from './templates'
-export type { ReadCheckCommentOpts, WrongMoveOpts, SolvedOpts } from './templates'
+export type {
+  ReadCheckCommentOpts,
+  WrongMoveOpts,
+  WrongDefenceOpts,
+  SolvedOpts,
+} from './templates'